WIP - reorganise code
This moves the todo and wikis items in their own modules.
