+++ self_wiki/__init__.py
@@ -7,6 +7,8 @@ from os.path import exists, expanduser, join as pjoin
 from flask import Flask
 from git import Repo
 
+from self_wiki.wiki import repository
+
 __version__ = "0.6.2"
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
@@ -28,21 +30,6 @@ logger.info("Using %s as content root", CONTENT_ROOT)
 if not exists(CONTENT_ROOT):
     os.mkdir(CONTENT_ROOT)
 
-MD_EXTS = [
-    "extra",
-    "admonition",
-    "codehilite",
-    "meta",
-    "sane_lists",
-    "smarty",
-    "toc",
-    "wikilinks",
-]
-logger.info("Enabled markdown extensions: %s", ", ".join(MD_EXTS))
-
-repository = None
 if exists(pjoin(CONTENT_ROOT, ".git")):
     repository = Repo(CONTENT_ROOT)
     logger.info("Git integration is enabled. self.wiki will commit changes")
-
-import self_wiki.views
