+++ self_wiki/views.py
@@ -1,4 +1,5 @@
 """Contains the flask views and their related objects."""
+import logging
 import os
 import re
 from datetime import date
@@ -13,160 +14,15 @@ from flask import (
 )
 from flask.views import MethodView
 
-from markdown import Markdown
+from self_wiki import CONTENT_ROOT, app, repository
+from self_wiki.todo import TodoList
+from self_wiki.utils import write_todo_to_journal
+from self_wiki.wiki import RecentFileManager, Page
 
-from self_wiki import CONTENT_ROOT, MD_EXTS, app, logger, repository
-from self_wiki.utils import RecentFileManager, TodoList
+logger = logging.getLogger(__name__)
 
 RECENT_FILES = RecentFileManager(CONTENT_ROOT)
-
-
-class Page:
-    """
-    Container for a markdown file.
-
-    Basically, all manipulation on .md files should go via this
-    """
-
-    @classmethod
-    def from_md(cls, markdown: str):
-        """
-        Create a Page object from some markdown content.
-
-        :param markdown:
-        :return:
-        """
-        bogus_page = Page("bogus_path")  # fixme, i'm ugly
-        bogus_page.markdown = markdown
-        bogus_page.path = bogus_page.title
-        return bogus_page
-
-    def __init__(self, path, level=0):
-        """
-        Create a new Page representation.
-
-        :param path: path this page's data on disk
-        :param level: How deep are we in the rabbit hole?
-        """
-        self.path = pjoin(CONTENT_ROOT, path)
-        self.level = level
-        if path[-3:] != ".md":
-            self.path = self.path + ".md"
-        self.markdown = ""
-        self.converter = Markdown(extensions=MD_EXTS, output_format="html5")
-        self.meta = None
-        self.subpages = []
-        self.load()
-
-    def load(self):
-        """
-        Load the markdown data from disk.
-
-        Also sets object properties according to filesystem state.
-        """
-        if not exists(self.path):
-            return
-        with open(self.path, "r") as markdown_file:
-            logger.debug(
-                "Found existing page content at %s. " "Loading at level %d",
-                self.path,
-                self.level,
-            )
-            self.markdown = markdown_file.read()
-
-        # We need a way to make sure we don't read an entire directory tree
-        if self.level > 0:
-            return
-        subpages_dir = self.path[:-3]  # remove the .md
-        if exists(subpages_dir) and isdir(subpages_dir):
-            for markdown_file in os.listdir(subpages_dir):
-                if isdir(markdown_file) or not markdown_file.endswith(".md"):
-                    continue
-                logger.debug(
-                    "Found child page %s at level %d",
-                    markdown_file,
-                    self.level,
-                )
-                self.subpages.append(
-                    Page(
-                        pjoin(subpages_dir, markdown_file),
-                        level=self.level + 1,
-                    )
-                )
-
-    def save(self):
-        """Persist the Page object on disk and update the recent files list."""
-        if os.path.sep in self.path and not exists(dirname(self.path)):
-            os.makedirs(dirname(self.path))
-        with open(self.path, "w+") as save_file:
-            save_file.write(self.markdown)
-        RECENT_FILES.update(self.path)
-        if repository is not None:
-            repository.index.add([self.path])
-            if repository.index.diff:
-                logger.info("Adding changes to page %s to git", self.title)
-                repository.index.commit(message="Change {}".format(self.title))
-
-    @property
-    def relpath(self):
-        """Return the page's path, relative to the configured content root."""
-        path = self.path[len(CONTENT_ROOT) - 1 :]  # noqa: E203
-        return path
-
-    @property
-    def title(self):
-        """
-        Return the title of the page.
-
-        This is computed either from the markdown's metadata
-        ('Title:' as one of the pages' header), or the first level 1 header, or
-        the pages' path
-        """
-        if not self.meta:
-            self.render()
-        if "title" in self.meta:
-            return self.meta["title"][0]
-        for line in self.markdown.split("\n"):
-            if line.startswith("# "):
-                return line[2:]
-        return self.relpath[:-3]
-
-    def render(self):
-        """Render the markdown to HTML, using the object's converter."""
-        html = self.converter.convert(self.markdown)
-        self.meta = self.converter.Meta  # pylint: disable=E1101
-        return html
-
-
 TODO_LIST = TodoList(pjoin(CONTENT_ROOT, "todos.json"))
-
-
-def write_todo_to_journal(todo: dict):
-    """
-    Write the object to a Page, denoted as journal in the URL.
-
-    The given item should have the following keys: 'id', 'text'
-    """
-    p = Page(date.today().strftime("journal/%Y/%m/%d"))
-    # load existing
-    p.load()
-    if p.markdown == "":
-        # we are freeeee
-        p.markdown = """# journal du {d}
-
-## Done
-
-* {id}: {text}
-""".format(
-            d=date.today().strftime("%Y/%m/%d"), **todo
-        )
-        p.save()
-        return
-    match = re.match(r"#+ *Done\n+", p.markdown)
-    if not match:
-        p.markdown = p.markdown + "\n\n## Done\n\n"
-    p.markdown = p.markdown + "* {id}: {text}\n".format(**todo)
-    p.save()
 
 
 class TodoView(MethodView):
@@ -215,9 +71,10 @@ def save(path):  # noqa: D103
     if not request.is_json:
         return 401
     markdown = request.json["markdown"]
-    page_to_save = Page(path)
+    page_to_save = Page(path, CONTENT_ROOT)
     page_to_save.markdown = markdown
     page_to_save.save()
+    RECENT_FILES.update(page_to_save.path)
     return "OK", 201
 
 
@@ -243,7 +100,7 @@ def upload(path):  # noqa: D103
 @app.route("/", defaults={"path": "index"}, methods=["DELETE"])
 @app.route("/<path:path>", methods=["DELETE"])
 def delete(path):  # noqa: D103
-    p = Page(path)
+    p = Page(path, CONTENT_ROOT)
     try:
         os.remove(p.path)
         RECENT_FILES.delete(p.path)
@@ -264,8 +121,8 @@ def edit(path):  # noqa: D103
     # Nooooon rien de rien....
     return render_template(
         "edit.html.j2",
-        page=Page(path),
-        recent=(Page(f["path"]) for f in RECENT_FILES.get(9)),
+        page=Page(path, CONTENT_ROOT),
+        recent=(Page(f["path"], CONTENT_ROOT) for f in RECENT_FILES.get(9)),
     )
 
 
@@ -282,11 +139,13 @@ def page(path):  # noqa: D103
         )
     if str(path).endswith("/"):
         return redirect(path[:-1])
-    page_to_view = Page(path)
+    page_to_view = Page(pjoin(CONTENT_ROOT, path))
     if page_to_view.markdown == "":
         return redirect(path + "/edit")
     return render_template(
         "page.html.j2",
         page=page_to_view,
-        recent=(Page(f["path"]) for f in RECENT_FILES.get(9)),
+        recent=(
+            Page(pjoin(CONTENT_ROOT, f["path"])) for f in RECENT_FILES.get(9)
+        ),
     )
