+++ tox.ini
@@ -5,6 +5,7 @@
 
 [tox]
 envlist = py35, py36, py37, pypy, black, flake8, mypy
+usedevelop = true
 
 [tox:jenkins]
 envlist = py35, py36, py37, pypy, flake8
